Split frontend route content into lazy-loaded modules
