+++ frontend/app/pages/TracksPage.tsx
@@ -1,4 +1,3 @@
-import { TrackRow } from '@/components/track/TrackRow'
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -7,40 +6,10 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from '@/components/ui/breadcrumb'
-import { Empty } from '@/components/ui/empty'
-import { Item, ItemContent, ItemMedia } from '@/components/ui/item'
-import { Skeleton } from '@/components/ui/skeleton'
-import { useData } from '@/contexts/DataContext'
 import loc from '@/lib/locales'
-import type { Track } from '@common/models/track'
-import type { ComponentProps } from 'react'
-import { twMerge } from 'tailwind-merge'
-import { PageHeader } from '../../components/PageHeader'
+import { TracksContent, type TracksContentProps } from './TracksContent'
 
-type TracksPageProps = { showLink?: true } & ComponentProps<'div'>
-
-function TrackRowList({ tracks }: Readonly<{ tracks: Track[] }>) {
-  if (tracks.length === 0) {
-    return (
-      <Empty className='border border-input text-sm text-muted-foreground'>
-        {loc.no.common.noItems}
-      </Empty>
-    )
-  }
-  return (
-    <div className='rounded-sm bg-background-secondary'>
-      {tracks.map(track => (
-        <TrackRow
-          key={track.id}
-          item={track}
-          className='py-3 first:pt-4 last:pb-4'
-        />
-      ))}
-    </div>
-  )
-}
-
-export default function TracksPage(props: Readonly<TracksPageProps>) {
+export default function TracksPage(props: Readonly<TracksContentProps>) {
   return (
     <div>
       <Breadcrumb>
@@ -55,64 +24,6 @@ export default function TracksPage(props: Readonly<TracksPageProps>) {
         </BreadcrumbList>
       </Breadcrumb>
       <TracksContent {...props} />
-    </div>
-  )
-}
-
-export function TracksContent({
-  className,
-  showLink,
-}: Readonly<TracksPageProps>) {
-  const { tracks, timeEntries, isLoadingData } = useData()
-
-  if (isLoadingData) {
-    return (
-      <div className={twMerge('flex flex-col', className)}>
-        <Item>
-          <ItemMedia>
-            <Skeleton className='size-8 rounded-sm' />
-          </ItemMedia>
-          <ItemContent>
-            <Skeleton className='h-6 w-24 rounded-sm' />
-            <Skeleton className='h-4 w-64 rounded-sm' />
-          </ItemContent>
-        </Item>
-
-        <div className='overflow-clip rounded-sm'>
-          <Skeleton className='h-16 w-full divide-y divide-border rounded-none' />
-          <Skeleton className='h-16 w-full divide-y divide-border rounded-none' />
-          <Skeleton className='h-16 w-full divide-y divide-border rounded-none' />
-        </div>
-      </div>
-    )
-  }
-
-  const tracksWithEntries = tracks.filter(t =>
-    timeEntries.some(te => te.track === t.id)
-  )
-
-  return (
-    <div className={twMerge('flex flex-col', className)}>
-      <PageHeader
-        title={loc.no.tracks.title}
-        description={loc.no.tracks.description}
-        to={showLink ? '/tracks' : undefined}
-        icon={'MapIcon'}
-      />
-
-      <TrackRowList
-        tracks={tracksWithEntries.filter(t => t.level !== 'custom')}
-      />
-
-      <PageHeader
-        title={loc.no.tracks.level.custom}
-        description={loc.no.tracks.customDescription}
-        icon={'WrenchIcon'}
-      />
-
-      <TrackRowList
-        tracks={tracksWithEntries.filter(t => t.level === 'custom')}
-      />
     </div>
   )
 }
